src/onecontainer_api/backends/mers-ffmpeg/requirements.txt
sanic==20.9.1
loguru==0.5.*
ffmpeg-python==0.2.0
sqlalchemy==1.3.23
cloudstore==0.2.0
netifaces==0.10.9